Safari-611.1.10/LayoutTests/http/tests/quicklook/rtf-document-domain-is-empty-string.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/ui-helper.js"></script>
<script src="resources/tap-run-test-hyperlink.js"></script>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFramesAsText();
}
</script>
</head>
<body>
<iframe src="resources/document-domain-is-empty-string.rtf" onload="runTest(this, 8 /* offsetInFrame */)"></iframe>
</body>
</html>
