Safari-611.1.10/LayoutTests/http/tests/quicklook/rtf-document-domain-is-empty-string-expected.txt



--------
Frame: '<!--frame1-->'
--------
PASS document.domain is .
